+++ packages/magic-entries-webpack-plugin/src/test/magic-entries-webpack-plugin.test.ts
@@ -1,3 +1,4 @@
+import 'setimmediate';
 import webpack, {Compiler, Configuration, Stats} from 'webpack';
 import VirtualModulesPlugin from 'webpack-virtual-modules';