+++ packages/react-server/src/webpack-plugin/test/webpack-plugin.test.ts
@@ -1,3 +1,5 @@
+import 'setimmediate';
+
 import path from 'path';
 
 import webpack, {Compiler} from 'webpack';