+++ packages/sewing-kit-plugin-quilt/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Changed
+
+- Update `@sewing-kit/*` packages [#1948](https://github.com/Shopify/quilt/pull/1948)
 
 ## 0.2.1 - 2021-06-17